+++ src/main/java/rip/shukaaa/enums/RowSlicerDirection.java
@@ -1,0 +1,5 @@
+package rip.shukaaa.enums;
+
+public enum RowSlicerDirection {
+    ROW, COLUMN
+}