+++ src/main/java/rip/shukaaa/image/effects/Transform.java
@@ -1,6 +1,7 @@
 package rip.shukaaa.image.effects;
 
 import rip.shukaaa.image.Pixel;
+import rip.shukaaa.image.ShukaaaImage;
 
 import java.util.ArrayList;
 
@@ -13,5 +14,17 @@ public class Transform {
             reversed.add(pixels.get(i));
         }
         return reversed;
+    }
+
+    public static void distortionFlip(ShukaaaImage image) {
+        ArrayList<Pixel> pixels = new ArrayList<>();
+        for (int x = 0; x < image.getHeight(); x++) {
+            for (int y = 0; y < image.getWidth(); y++) {
+                pixels.add(new Pixel(image.getRGB(y, x)));
+            }
+        }
+
+        pixels = flip(pixels);
+        image.setPixels(pixels);
     }
 }