+++ changelog.md
@@ -1,3 +1,8 @@
+### 1.2
+- Added Noise Shuffler Effect
+- Added Distortion Flip Effect
+- Scale down too large images for the view so the window doesn't get too large
+
 ### 1.1
 - Added Transform Menu Item
     - Added Flip